Create a Rails-like structure for turbo stream custom methods customization
