+++ app/javascript/application.js
@@ -1,14 +1,14 @@
 // Configure your import map in config/importmap.rb. Read more: https://github.com/rails/importmap-rails
 import "@hotwired/turbo-rails"
 import "controllers"
+import "initializers"
 
 import "@fortawesome/fontawesome-free"
 import "channels"
 
 import jquery from 'jquery';
 import select2 from 'select2';
-import dispatcher from "./services/dispatcher";
-import "./services/turbo_stream_dispatcher";
+import dispatcher from "services/dispatcher";
 
 window.jQuery = jquery
 window.$ = jquery
