+++ app/javascript/initializers/index.js
@@ -1,0 +1,1 @@
+import "initializers/turbo_stream"
