+++ app/javascript/initializers/turbo_stream/index.js
@@ -1,0 +1,1 @@
+import "initializers/turbo_stream/dispatch_event"
