+++ config/importmap.rb
@@ -21,6 +21,7 @@ pin "just-extend" # @5.1.1 required by dropzone
 pin "fuse.js" # @7.1.0
 
 pin_all_from "app/javascript/controllers", under: "controllers"
+pin_all_from "app/javascript/initializers", under: "initializers"
 pin_all_from "app/javascript/controllers/visualizations", under: "controllers"
 pin_all_from "app/javascript/services", under: "services"
 pin "@rails/request.js", to: "@rails--request.js.js" # @0.0.11
